Add Franca FSDP (beta) support and training stability fixes
Changed:
  - Add --distributed-mode fsdp support to train_franca with two FSDP roots (student + teacher)

Fixed:
  - Freeze EMA teachers earlier in train_dino_v2, train_dino_v2_dist, train_franca and train_capi to avoid optimizer-state mismatches on resume
  - Cast pred and target to float32 in NEPA prediction_loss so normalization and cosine similarity run in full precision under autocast
  - Use round() instead of int() in SmoothedValue count reduction to avoid off-by-one from ReduceOp.AVG floating-point imprecision

+++ birder/net/ssl/nepa.py
@@ -26,8 +26,8 @@ def prediction_loss(pred: torch.Tensor, target: torch.Tensor, shift: bool = True
 
     target = target.detach()
 
-    pred = F.normalize(pred, dim=-1)
-    target = F.normalize(target, dim=-1)
+    pred = F.normalize(pred.float(), dim=-1)
+    target = F.normalize(target.float(), dim=-1)
 
     return -(pred * target).sum(dim=-1).mean()
